requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
drf-yasg   # or drf-spectacular
gunicorn
python-dotenv
